+++ babel.config.js
@@ -1,13 +1,10 @@
 module.exports = {
-    presets: [
-        [
-            '@babel/preset-env',
-            {
-                targets: {
-                    node: 'current',
-                    edge: '18'
-                },
-            },
-        ],
-    ],
+	presets: [
+		[
+			'@babel/preset-env',
+			{
+				targets: 'defaults',
+			},
+		],
+	],
 };